Cleaning up a little
