+++ grunt/aliases.js
@@ -1,0 +1,5 @@
+// the task aliases
+module.exports = {
+  'default': [],
+  'debug': []
+}
