+++ grunt/config.js
@@ -1,0 +1,5 @@
+// initial grunt config. register-module, appends onto this with tasks
+module.exports = {
+  concat: {},
+  less: {}
+};
